+++ server/__init__.py
@@ -1,0 +1,3 @@
+"""
+Virtual Model Emulator server package
+"""